requirements.txt
boto3>=1.26.0
strands-agents
bedrock-agentcore
bedrock-agentcore-starter-toolkit
